+++ app/frontend/src/components/forestrysectoropscreening/Form.vue
@@ -1,16 +1,17 @@
 <template>
   <div>
-    <v-stepper v-model="step" alt-labels class="silv-stepper elevation-0" @change="setStep">
+    <Landing v-if="step < 1" />
+    <v-stepper v-else v-model="step" alt-labels class="silv-stepper elevation-0" @change="setStep">
       <v-container v-if="!submissionComplete" fluid class="pa-0">
         <v-row class="header-row" no-gutters>
-          <v-col cols="12" xl="10" offset-xl="1">
+          <v-col cols="12" xl="8" offset-xl="2">
             <v-stepper-header class="elevation-0 hidden-xs-only">
               <v-stepper-step
                 :complete="step > 1"
                 edit-icon="check"
                 :editable="step > 1"
                 step="1"
-              >Before You Begin</v-stepper-step>
+              >Contact Information</v-stepper-step>
 
               <v-divider></v-divider>
 
@@ -19,7 +20,7 @@
                 edit-icon="check"
                 :editable="step > 2"
                 step="2"
-              >Contact Information</v-stepper-step>
+              >Before Operations Begin</v-stepper-step>
 
               <v-divider></v-divider>
 
@@ -28,7 +29,7 @@
                 edit-icon="check"
                 :editable="step > 3"
                 step="3"
-              >Before Operations Begin</v-stepper-step>
+              >After Workers Arrive</v-stepper-step>
 
               <v-divider></v-divider>
 
@@ -37,20 +38,11 @@
                 edit-icon="check"
                 :editable="step > 4"
                 step="4"
-              >After Workers Arrive</v-stepper-step>
-
-              <v-divider></v-divider>
-
-              <v-stepper-step
-                :complete="step > 5"
-                edit-icon="check"
-                :editable="step > 5"
-                step="5"
               >If Workers Become Ill</v-stepper-step>
 
               <v-divider></v-divider>
 
-              <v-stepper-step step="6">Review</v-stepper-step>
+              <v-stepper-step step="5">Review</v-stepper-step>
             </v-stepper-header>
           </v-col>
         </v-row>
@@ -79,10 +71,6 @@
               <v-stepper-content step="5">
                 <Step5 />
               </v-stepper-content>
-
-              <v-stepper-content step="6">
-                <Step6 />
-              </v-stepper-content>
             </v-stepper-items>
           </v-col>
         </v-row>
@@ -93,33 +81,32 @@
 
 <script>
 import { mapGetters, mapMutations } from 'vuex';
-
-import Step1 from '@/components/minesoperatorscreening/Step1.vue';
-import Step2 from '@/components/minesoperatorscreening/Step2.vue';
-import Step3 from '@/components/minesoperatorscreening/Step3.vue';
-import Step4 from '@/components/minesoperatorscreening/Step4.vue';
-import Step5 from '@/components/minesoperatorscreening/Step5.vue';
-import Step6 from '@/components/minesoperatorscreening/Step6.vue';
+import Landing from '@/components/forestrysectoropscreening/Landing.vue';
+import Step1 from '@/components/forestrysectoropscreening/Step1.vue';
+import Step2 from '@/components/forestrysectoropscreening/Step2.vue';
+import Step3 from '@/components/forestrysectoropscreening/Step3.vue';
+import Step4 from '@/components/forestrysectoropscreening/Step4.vue';
+import Step5 from '@/components/forestrysectoropscreening/Step5.vue';
 
 export default {
-  name: 'MinesAttestationForm',
+  name: 'ForestrySectorAttestationForm',
   components: {
+    Landing,
     Step1,
     Step2,
     Step3,
     Step4,
-    Step5,
-    Step6
+    Step5
   },
   data () {
     return {
     };
   },
   computed: {
-    ...mapGetters('form', ['step', 'submissionComplete']),
+    ...mapGetters('forestrySectorOpScreeningForm', ['step', 'submissionComplete']),
   },
   methods: {
-    ...mapMutations('form', ['setStep']),
+    ...mapMutations('forestrySectorOpScreeningForm', ['setStep']),
   }
 };
 </script>
